Make api less verbose; fix <Plug> maps
This has the side effect of making `<Plug>Occurrence*` execute faster
because nvim does not have to wait to differentiate between keymaps that
share the same prefix.

e.g., previously `<Plug>OccurrenceNext` would appear to be slower than
`<Plug>OccurrenceNextMatch` because nvim would wait `:h timeoutlen` to
disambiguate the two in the first case, but not in the second.

Making the API a bit less verbose and updating <Plug>* mappings to be
<Plug>(Occurrence*) instead mitigates this.